Move code into tests/test_base.py, remove params_set
