+++ tests/fenics/test_base.py
@@ -1,36 +1,30 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-from ..test_base import *
-
 from fenics import *
 from tlm_adjoint.fenics import *
 from tlm_adjoint.fenics import manager as _manager
 from tlm_adjoint.fenics.backend import backend_Constant, backend_Function
-from tlm_adjoint.fenics.backend_code_generator_interface import \
-    complex_mode, interpolate_expression
+from tlm_adjoint.fenics.backend_code_generator_interface import (
+    complex_mode, interpolate_expression)
 from tlm_adjoint.alias import gc_disabled
 from tlm_adjoint.override import override_method
 
-import copy
-import functools
+from ..test_base import chdir_tmp_path, seed_test, tmp_path
+from ..test_base import run_example as _run_example
+
 import gc
-import hashlib
-import inspect
 import logging
 import mpi4py.MPI as MPI
-import numpy as np
 try:
     from operator import call
 except ImportError:
     # For Python < 3.11, following Python 3.11 API
     def call(obj, /, *args, **kwargs):
         return obj(*args, **kwargs)
-from operator import itemgetter
 import os
 import petsc4py.PETSc as PETSc
 import pytest
-import runpy
 import weakref
 
 __all__ = \
@@ -38,6 +32,7 @@ __all__ = \
         "complex_mode",
         "interpolate_expression",
 
+        "chdir_tmp_path",
         "run_example",
         "seed_test",
         "setup_test",
@@ -87,43 +82,14 @@ def setup_test():
         gc.enable()
 
 
-def seed_test(fn):
-    @functools.wraps(fn)
-    def wrapped_fn(*args, **kwargs):
-        h_kwargs = dict(kwargs)
-        if "tmp_path" in inspect.signature(fn).parameters:
-            # Raises an error if tmp_path is a positional argument
-            del h_kwargs["tmp_path"]
-
-        h = hashlib.sha256()
-        h.update(fn.__name__.encode("utf-8"))
-        h.update(str(args).encode("utf-8"))
-        h.update(str(sorted(h_kwargs.items(), key=itemgetter(0))).encode("utf-8"))  # noqa: E501
-        seed = int(h.hexdigest(), 16) + MPI.COMM_WORLD.rank
-        seed %= 2 ** 32
-        np.random.seed(seed)
-
-        return fn(*args, **kwargs)
-    return wrapped_fn
-
-
-def params_set(names, *values):
-    if len(values) > 1:
-        sub_params = params_set(names[1:], *values[1:])
-        params = []
-        for value in values[0]:
-            for sub_params_ in sub_params:
-                new_params = copy.deepcopy(sub_params_)
-                new_params[names[0]] = value
-                params.append(new_params)
-    else:
-        params = [{names[0]:value} for value in values[0]]
-    return params
-
-
-@pytest.fixture(params=params_set(["enable_caching", "defer_adjoint_assembly"],
-                                  [True, False],
-                                  [True, False]))
+@pytest.fixture(params=[{"enable_caching": True,
+                         "defer_adjoint_assembly": True},
+                        {"enable_caching": True,
+                         "defer_adjoint_assembly": False},
+                        {"enable_caching": False,
+                         "defer_adjoint_assembly": True},
+                        {"enable_caching": False,
+                         "defer_adjoint_assembly": False}])
 def test_configurations(request):
     parameters["tlm_adjoint"]["EquationSolver"]["enable_jacobian_caching"] \
         = request.param["enable_caching"]
@@ -205,13 +171,6 @@ def test_leaks():
     manager.reset("memory", {"drop_references": False})
 
 
-@pytest.fixture
-def tmp_path(tmp_path):
-    if MPI.COMM_WORLD.rank != 0:
-        tmp_path = None
-    return MPI.COMM_WORLD.bcast(tmp_path, root=0)
-
-
 def run_example(example, *,
                 add_example_path=True, clear_forward_globals=True):
     if add_example_path:
@@ -220,14 +179,7 @@ def run_example(example, *,
                                 "examples", "fenics", example)
     else:
         filename = example
-
-    start_manager()
-    gl = runpy.run_path(filename)
-
-    if clear_forward_globals:
-        # Clear objects created by the script. Requires the script to define a
-        # 'forward' function.
-        gl["forward"].__globals__.clear()
+    _run_example(filename, clear_forward_globals=clear_forward_globals)
 
 
 ls_parameters_cg = {"linear_solver": "cg",
