+++ tests/firedrake/test_base.py
@@ -1,8 +1,6 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-from ..test_base import *
-
 from firedrake import *
 from tlm_adjoint.firedrake import *
 from tlm_adjoint.firedrake import manager as _manager
@@ -12,25 +10,20 @@ from tlm_adjoint.firedrake.backend_code_generator_interface import (
 from tlm_adjoint.alias import gc_disabled
 from tlm_adjoint.override import override_method
 
-import copy
-import functools
+from ..test_base import chdir_tmp_path, seed_test, tmp_path
+from ..test_base import (
+    run_example as _run_example, run_example_notebook as _run_example_notebook)
+
 import gc
-import hashlib
-import inspect
-import json
 import logging
-import mpi4py.MPI as MPI
-import numpy as np
 try:
     from operator import call
 except ImportError:
     # For Python < 3.11, following Python 3.11 API
     def call(obj, /, *args, **kwargs):
         return obj(*args, **kwargs)
-from operator import itemgetter
 import os
 import pytest
-import runpy
 import sys
 import weakref
 
@@ -39,6 +32,7 @@ __all__ = \
         "complex_mode",
         "interpolate_expression",
 
+        "chdir_tmp_path",
         "run_example",
         "run_example_notebook",
         "seed_test",
@@ -85,43 +79,14 @@ def setup_test():
     clear_caches()
 
 
-def seed_test(fn):
-    @functools.wraps(fn)
-    def wrapped_fn(*args, **kwargs):
-        h_kwargs = dict(kwargs)
-        if "tmp_path" in inspect.signature(fn).parameters:
-            # Raises an error if tmp_path is a positional argument
-            del h_kwargs["tmp_path"]
-
-        h = hashlib.sha256()
-        h.update(fn.__name__.encode("utf-8"))
-        h.update(str(args).encode("utf-8"))
-        h.update(str(sorted(h_kwargs.items(), key=itemgetter(0))).encode("utf-8"))  # noqa: E501
-        seed = int(h.hexdigest(), 16) + MPI.COMM_WORLD.rank
-        seed %= 2 ** 32
-        np.random.seed(seed)
-
-        return fn(*args, **kwargs)
-    return wrapped_fn
-
-
-def params_set(names, *values):
-    if len(values) > 1:
-        sub_params = params_set(names[1:], *values[1:])
-        params = []
-        for value in values[0]:
-            for sub_params_ in sub_params:
-                new_params = copy.deepcopy(sub_params_)
-                new_params[names[0]] = value
-                params.append(new_params)
-    else:
-        params = [{names[0]:value} for value in values[0]]
-    return params
-
-
-@pytest.fixture(params=params_set(["enable_caching", "defer_adjoint_assembly"],
-                                  [True, False],
-                                  [True, False]))
+@pytest.fixture(params=[{"enable_caching": True,
+                         "defer_adjoint_assembly": True},
+                        {"enable_caching": True,
+                         "defer_adjoint_assembly": False},
+                        {"enable_caching": False,
+                         "defer_adjoint_assembly": True},
+                        {"enable_caching": False,
+                         "defer_adjoint_assembly": False}])
 def test_configurations(request):
     parameters["tlm_adjoint"]["EquationSolver"]["enable_jacobian_caching"] \
         = request.param["enable_caching"]
@@ -196,13 +161,6 @@ def test_leaks():
     manager.reset("memory", {"drop_references": False})
 
 
-@pytest.fixture
-def tmp_path(tmp_path):
-    if MPI.COMM_WORLD.rank != 0:
-        tmp_path = None
-    return MPI.COMM_WORLD.bcast(tmp_path, root=0)
-
-
 def run_example(example, *,
                 add_example_path=True, clear_forward_globals=True):
     if add_example_path:
@@ -211,43 +169,18 @@ def run_example(example, *,
                                 "examples", "firedrake", example)
     else:
         filename = example
-
-    start_manager()
-    gl = runpy.run_path(filename)
-
-    if clear_forward_globals:
-        # Clear objects created by the script. Requires the script to define a
-        # 'forward' function.
-        gl["forward"].__globals__.clear()
+    _run_example(filename, clear_forward_globals=clear_forward_globals)
 
 
 def run_example_notebook(example, tmp_path, *,
                          add_example_path=True):
-    if MPI.COMM_WORLD.size > 1:
-        raise RuntimeError("Serial only")
-
     if add_example_path:
         filename = os.path.join(os.path.dirname(__file__),
                                 os.path.pardir, os.path.pardir,
                                 "docs", "source", "examples", example)
     else:
         filename = example
-    tmp_filename = os.path.join(tmp_path, "tmp.py")
-
-    with open(filename, "r") as nb_h, open(tmp_filename, "w") as py_h:
-        nb = json.load(nb_h)
-        if nb["metadata"]["language_info"]["name"] != "python":
-            raise RuntimeError("Expected a Python notebook")
-
-        for cell in nb["cells"]:
-            if cell["cell_type"] == "code":
-                for line in cell["source"]:
-                    py_h.write(line)
-                py_h.write("\n\n")
-
-    reset_manager("memory", {"drop_references": False})
-    run_example(tmp_filename,
-                add_example_path=False, clear_forward_globals=False)
+    _run_example_notebook(filename, tmp_path)
 
 
 ls_parameters_cg = {"ksp_type": "cg",
