+++ tests/firedrake/test_examples.py
@@ -40,6 +40,7 @@ def test_0_getting_started(setup_test, tmp_path):
 @pytest.mark.skipif(complex_mode, reason="real only")
 @pytest.mark.skipif(MPI.COMM_WORLD.size > 1, reason="serial only")
 def test_1_time_independent(setup_test, tmp_path):
+    reset_manager("memory", {"drop_references": False})
     run_example_notebook("1_time_independent.ipynb", tmp_path)
 
 
@@ -48,6 +49,7 @@ def test_1_time_independent(setup_test, tmp_path):
 @pytest.mark.skipif(complex_mode, reason="real only")
 @pytest.mark.skipif(MPI.COMM_WORLD.size > 1, reason="serial only")
 def test_2_verification(setup_test, tmp_path):
+    reset_manager("memory", {"drop_references": False})
     run_example_notebook("2_verification.ipynb", tmp_path)
